requirements.txt
streamlit
inference
opencv-python-headless
pillow